refactor: merge package adSense into google_ads
